fix: update health points on server chat

+++ src/main/java/MainGameStage/GameTimer.java
@@ -478,7 +478,7 @@ class GameTimer extends AnimationTimer{
 				player.setHealth();
 				player.setIsAlive();
 				if (player.getName().equals(this.username) && player.getIsAlive())
-					connection.send("SERVER_CHAT " + username + " has " + player.getHealth() + " remaining health.");
+					connection.send("SERVER_CHAT " + username + " has " + ((int) player.getHealth()) + " remaining health.");
 				break;
 			}
 		}
